Update migration scripts and schema version
